Refactored code to expose implementation
Instead of interacting with a wrapper around counterfactual
implementations, the user instantiates an object that specifies the
implementation directly. Specifically:

- The user instantiates the WachterCounterfactual class, specifying the
  predictor, its type, loss and method specifications, the feature range
  and the framework they wish to use for optimization (should be the
  same as the framework the predictor is written in for whitebox
  predictors). Only the predictor argument is mandatory.
- The WachterCounterfactual class is a wrapper around the actual
  implementation, its main role is to implement the Alibi API
- The _WachterCounterfactual private class which implements the search
  process and is the _explainer attribute of the public namesake. This
  allows batch support (and distributed explanations) in Alibi

The package alibi.explainers.base has been removed.
