--- alibi/explainers/base/__init__.py
@@ -1,8 +1,0 @@
-from .registry import register_explainer, explainer_registry
-from .loaders import get_implementation
-
-__all__ = [
-    "register_explainer",
-    "explainer_registry",
-    "get_implementation",
-]
